cisco_routers/requirements.txt
django>2
djangorestframework==3.8.0
django-import-export
django-macaddress
